# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/company/appeal": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "公司进行申诉",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company"
                ],
                "summary": "公司进行申诉",
                "parameters": [
                    {
                        "description": "公司申诉信息",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CompanyAppealDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/company/history": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "返回所有的企业申诉信息给政府视图",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company"
                ],
                "summary": "返回所有的企业申诉信息给政府视图",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AppealList"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.FailureResponse"
                        }
                    }
                }
            }
        },
...
        "/gvmt/changeAccusStatus": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "修改待审核的排污实地图列表状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gvmt(admin)"
                ],
                "summary": "修改待审核的排污实地图列表状态",
                "parameters": [
                    {
                        "description": "status=1表示不通过,status=2表示等待,status=3表示通过",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AccusStatusDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/gvmt/changeAplStatus": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "修改申诉列表状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gvmt(admin)"
                ],
                "summary": "修改申诉列表状态",
                "parameters": [
                    {
                        "description": "status=1表示不通过,status=2表示等待,status=3表示通过",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AplStatusDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    }
                }
            }
        },
        "/gvmt/changeStatus": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "政府修改企业状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gvmt(admin)"
                ],
                "summary": "政府修改企业状态",
                "parameters": [
                    {
                        "description": "修改公司状态为status, 1为正常, 2为可疑违规",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.Company"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    }
                }
            }
        },
...
        "/gvmt/getAppealList": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "返回所有的企业申诉信息给政府视图",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gvmt(admin)"
                ],
                "summary": "返回所有的企业申诉信息给政府视图",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AppealList"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.FailureResponse"
                        }
                    }
                }
            }
        },
...
        "/visitor/history": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "返回所有的企业申诉信息给政府视图",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visitor"
                ],
                "summary": "返回所有的企业申诉信息给政府视图",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.AccusationList"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.FailureResponse"
                        }
                    }
                }
            }
        },
        "/visitor/upload": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "进行举报",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "visitor"
                ],
                "summary": "进行举报",
                "parameters": [
                    {
                        "description": "举报信息",
                        "name": "object",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.UserPostDto"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.SuccessResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/models.FailureResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "models.AccusStatusDto": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "9cade01e-4c66-4c57-910a-2efe5f7e3dff"
                },
                "status": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.Accusation": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "string",
                    "example": "17280000089799"
                },
                "id": {
                    "description": "文章uuid",
                    "type": "string",
                    "example": "a05cde18-488b-4dc4-afa1-df9e0ccf16a1"
                },
                "pic": {
                    "description": "帖子图片路径",
                    "type": "string",
                    "example": "图片内容"
                },
                "post": {
                    "description": "帖子内容",
                    "type": "string",
                    "example": "文字内容"
                },
                "status": {
                    "description": "举报状态",
                    "type": "integer",
                    "example": 2
                },
                "time": {
                    "description": "时间",
                    "type": "string",
                    "example": "2022-01-28T23:57:30+08:00"
                },
                "userId": {
                    "description": "用户名",
                    "type": "string",
                    "example": "123123"
                }
            }
        },
        "models.AccusationList": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "200"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Accusation"
                    }
                },
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "models.AplStatusDto": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "967d7ae4-1c39-40f8-befd-222d768d70f7"
                },
                "status": {
                    "type": "integer",
                    "example": 1
                }
            }
        },
        "models.Appeal": {
            "type": "object",
            "properties": {
                "company_id": {
                    "type": "string",
                    "example": "17280000089799"
                },
                "id": {
                    "description": "文章uuid",
                    "type": "string",
                    "example": "967d7ae4-1c39-40f8-befd-222d768d70f7"
                },
                "pic": {
                    "description": "帖子图片路径",
                    "type": "string",
                    "example": "图片内容"
                },
                "post": {
                    "description": "帖子内容",
                    "type": "string",
                    "example": "文字内容"
                },
                "status": {
                    "description": "举报状态",
                    "type": "integer",
                    "example": 2
                },
                "time": {
                    "description": "时间",
                    "type": "string",
                    "example": "2022-01-28T23:45:54+08:00"
                }
            }
        },
        "models.AppealList": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "200"
                },
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Appeal"
                    }
                },
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "models.Company": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "172800010800110000"
                },
                "status": {
                    "type": "integer",
                    "example": 2
                }
            }
        },
        "models.CompanyAppealDto": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "17280000089799"
                },
                "pic": {
                    "type": "string",
                    "example": "图片内容"
                },
                "post": {
                    "type": "string",
                    "example": "文字申诉内容"
                }
            }
        },
...
        "models.FailureResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "401"
                },
                "msg": {
                    "type": "string",
                    "example": "token invalid"
                }
            }
        },
...
        "models.SuccessResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "200"
                },
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "models.UserPostDto": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "example": "17280000089799"
                },
                "pic": {
                    "type": "string",
                    "example": "图片内容"
                },
                "post": {
                    "type": "string",
                    "example": "文字内容"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}